Implemented MBC2 and MBC5. Change uint8_t value of some registers like TIMAClockCounter and selectedROM/RAMBank and its functions.

+++ include/Timer.h
@@ -26,7 +26,7 @@ private:
     IOHandler * ioHandler;
 
     uint16_t divCycles;
-    uint8_t TIMACycles;
+    uint16_t TIMACycles;
     uint16_t frequency;
 };
